Made a lot of changes...
Canceling is now done by SynchronizationProcessDialog and isCancelled()
is queried.

messages may now be delayed to reduce the number of invocations of
Swingutilities.invokeLater(Runnable).

Status updates may now delayed to reduce the number of invocations of
SwingUtilities.invokeLater(Runnable).

The progress bar is now being updated by another thread. The time to
count down from is being delegated.

"ignore modification time when file is equal" is purged the new strategy
is "elastic time comparison". If the files in the target directory
become younger by one second, as sometimes is the case, the time
comparison can be configured to allow a modification time +/- 1 second
for the file in the target directory from the time stored in the
database.

improved the integrity check.

changed while(iterator.hasNext()) loops to for each loops.

changed the strategy when encountering invalid files to inform user and
abort.

in unidirectional mode a file that is missing in the target directory is
now being copied again.

there is a second delete empty directories segment now after the delete
files segment that now also searches for empty directories as they
appear.

Statistics to project the process duration for the progress bar are
being collected now.
* 1 Either quick or deep analysis average duration and count.
* 2 data copy segment average duration, average megabytes, count and
total files copied count.
These statistics are saved with the operation data in sync.json.

When the method copy(CopyAction, OnlineDB, long, long) fails, the file
is now being skipped instead of aborting the whole process.

Added utility methods for communications with the EDT.

commented.